move queries over to dbaccessors
