+++ corehq/apps/es/users.py
@@ -142,14 +142,3 @@ def location(location_id):
             filters.term('domain_memberships.assigned_location_ids', location_id)
         ),
     )
-
-
-def user_ids_at_locations_and_descendants(location_ids):
-    return UserES().users_at_locations_and_descendants(location_ids).exclude_source().run().hits
-
-
-def user_ids_at_locations(location_ids):
-    return UserES().users_at_locations(location_ids).exclude_source().run().hits
-
-def user_ids_at_accessible_locations(domain_name, user):
-    return UserES().users_at_accessible_locations(domain_name, user).exclude_source().run().hits
