+++ corehq/apps/locations/dbaccessors.py
@@ -76,3 +76,15 @@ def get_users_location_ids(domain, user_ids):
               .run())
     location_ids = [r['assigned_location_ids'] for r in result.hits if 'assigned_location_ids' in r]
     return list(chain(*location_ids))
+
+
+def user_ids_at_locations_and_descendants(location_ids):
+    return UserES().users_at_locations_and_descendants(location_ids).exclude_source().run().hits
+
+
+def user_ids_at_locations(location_ids):
+    return UserES().users_at_locations(location_ids).exclude_source().run().hits
+
+
+def user_ids_at_accessible_locations(domain_name, user):
+    return UserES().users_at_accessible_locations(domain_name, user).exclude_source().run().hits
