+++ corehq/apps/reports/filters/users.py
@@ -4,12 +4,11 @@ from django.utils.translation import ugettext as _
 
 from corehq.apps.es import users as user_es, filters
 from corehq.apps.domain.models import Domain
-from corehq.apps.es.users import user_ids_at_locations_and_descendants
 from corehq.apps.groups.models import Group
+from corehq.apps.locations.dbaccessors import user_ids_at_locations_and_descendants
 from corehq.apps.reports.util import namedtupledict
 from corehq.apps.users.cases import get_wrapped_owner
 from corehq.apps.users.models import CommCareUser, WebUser
-from corehq.apps.es.users import user_ids_at_locations_and_descendants
 from corehq.util import remove_dups, flatten_list
 from dimagi.utils.decorators.memoized import memoized
 from corehq.apps.commtrack.models import SQLLocation
